requirements.txt
flask==3.0.2
requests==2.31.0
icalendar==5.0.11
python-dotenv==1.0.1
gunicorn==21.2.0
flask-cors==4.0.0
flask-session==0.6.0 